# apps/backend/docs/swagger.yaml
basePath: /
definitions:
...
  internal_handlers.AddMemberRequest:
    properties:
      email:
        type: string
      role:
        type: string
    type: object
...
  internal_handlers.CreateOrganizationRequest:
    properties:
      name:
        type: string
    type: object
...
  internal_handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_handlers.OrganizationMemberResponse:
    properties:
      avatar:
        type: string
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      name:
        type: string
      organizationId:
        type: string
      role:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  internal_handlers.OrganizationResponse:
    properties:
      adminId:
        type: string
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  internal_handlers.UpdateMemberRoleRequest:
    properties:
      role:
        type: string
    type: object
  internal_handlers.UpdateOrganizationRequest:
    properties:
      name:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  title: My API Name
  version: "1.0"
paths:
...
  /api/organizations:
    get:
      description: Lists all organizations where the user is a member
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.OrganizationResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: List user's organizations
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: Creates a new organization with the authenticated user as admin
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organization request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.OrganizationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Create a new organization
      tags:
      - organizations
  /api/organizations/{id}:
    get:
      description: Retrieves details of a specific organization
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.OrganizationResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Get organization details
      tags:
      - organizations
    put:
      consumes:
      - application/json
      description: Updates organization details (admin only)
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.OrganizationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Update organization
      tags:
      - organizations
  /api/organizations/{id}/members:
    get:
      description: Lists all members of an organization
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.OrganizationMemberResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: List organization members
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: Adds a user to an organization by email (admin only)
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Add member request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.AddMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.OrganizationMemberResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Add member to organization
      tags:
      - organizations
  /api/organizations/{id}/members/{memberId}:
    delete:
      description: Removes a user from an organization (admin only)
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Member ID / User ID
        in: path
        name: memberId
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Remove member from organization
      tags:
      - organizations
  /api/organizations/{id}/members/{memberId}/role:
    put:
      consumes:
      - application/json
      description: Updates a member's role in an organization (admin only)
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Member ID / User ID
        in: path
        name: memberId
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update role request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateMemberRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.OrganizationMemberResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Update member role
      tags:
      - organizations
...
swagger: "2.0"